Minor Polkadot api refactoring
