+++ src/helpers/Api.tsx
@@ -1,10 +1,69 @@
 import React, { useMemo, useState, useEffect, useContext } from 'react';
-import { ApiPromise } from '@polkadot/api';
+import { ApiPromise, WsProvider } from '@polkadot/api';
 import { ApiContext, ApiProps } from './ApiContext';
-import { getApi, PlasmNetwork } from './plasmUtils';
+import { PlasmNetwork } from './plasmUtils';
+import { plasmDefinitions, dustyDefinitions } from '@plasm/types';
 
 const DEFAULT_NETWORK = PlasmNetwork.Local;
 let api: ApiPromise;
+
+/**
+ * gets endpoint url for a given network
+ * @param network the network
+ */
+export function getNetworkEndpoint(network?: PlasmNetwork) {
+    let endpoint: string;
+
+    switch (network) {
+        case PlasmNetwork.Local:
+            endpoint = 'ws://127.0.0.1:9944';
+            break;
+        case PlasmNetwork.Dusty:
+            endpoint = 'wss://rpc.dusty.plasmnet.io/';
+            break;
+        case PlasmNetwork.Main: // main net endpoint will be the default value
+        default:
+            endpoint = 'wss://rpc.plasmnet.io';
+            break;
+    }
+
+    return endpoint;
+}
+
+/**
+ * creates ApiPromise for a given network
+ * @param network end point for the client to connect to
+ */
+export function getApi(network: PlasmNetwork): ApiPromise {
+    const types = network === PlasmNetwork.Main ? plasmDefinitions : dustyDefinitions;
+    const url = getNetworkEndpoint(network);
+    const provider = new WsProvider(url);
+    return new ApiPromise({
+        provider,
+        types: {
+            ...types,
+            // aliases that don't do well as part of interfaces
+            'voting::VoteType': 'VoteType',
+            'voting::TallyType': 'TallyType',
+            // chain-specific overrides
+            Address: 'GenericAddress',
+            Keys: 'SessionKeys4',
+            StakingLedger: 'StakingLedgerTo223',
+            Votes: 'VotesTo230',
+            ReferendumInfo: 'ReferendumInfoTo239',
+        },
+    });
+}
+
+/**
+ * establishes a connection between the client and the plasm node with the given endpoint.
+ * this will default to the main net node
+ * @param network end point for the client to connect to
+ */
+export async function createPlasmInstance(network?: PlasmNetwork) {
+    const api = getApi(network || PlasmNetwork.Main);
+    return await api.isReady;
+}
 
 function Api({ network = DEFAULT_NETWORK, children }: Props): React.ReactElement<Props> {
     const [isReady, setIsReady] = useState<boolean>(false);
